perf(moe): 中间量 f32→bf16 (Fp4Gemm 输出类型模板化) → prefill 871→893 tok/s
Fp4Gemm 输出从硬编码 f32 改为 template<OutT> (float→CUDA_R_32F / uint16_t→
CUDA_R_16BF, f32 累加)。MoE gu_out(SwiGLU 主读)/dn_out(Scatter 主读)f32→bf16,
两者主读减半。LtPlanKey 加 out_bits 判别位(q4t_tests 里 f32 直调 + bf16 MoE
路径同 M,N,K 会撞 plan cache)。

FP4 模型里 bf16 中间量精度损失远小于后续 FP4 量化, 风险低。全 68 测试通过零
警告: f32 直调路径(moe_load/fp4)逐位不变, bf16 MoE 路径(moe_gemm)max_rel
0.0077 < 容差。顺带修 lt_cache.cpp 既有 C++23 扩展 warning。

prefill 2560 tok 871→893 tok/s(+2.5%); MoE 本 session 累计 826→893(+8.1%)。

+++ include/q4t/quant/moe_gemm.h
@@ -34,8 +34,8 @@ namespace quant {
 struct MoEWorkspace {
   size_t a_packed_bytes = 0;  // [M*k, hs/2] e2m1
   size_t a_sf_bytes = 0;  // swizzled e4m3 (padded to 128-row atoms)
-  size_t gu_out_bytes = 0;  // [M*k, 2*moe_is] f32 gate/up GEMM output
-  size_t dn_out_bytes = 0;  // [M*k, hs] f32 down GEMM output
+  size_t gu_out_bytes = 0;  // [M*k, 2*moe_is] bf16 gate/up GEMM output
+  size_t dn_out_bytes = 0;  // [M*k, hs] bf16 down GEMM output
 
   size_t TotalBytes() const {
     return a_packed_bytes + a_sf_bytes + gu_out_bytes + dn_out_bytes;
@@ -46,8 +46,8 @@ struct MoEWorkspace {
 
   uint8_t* a_packed;
   uint8_t* a_sf;
-  float* gu_out;
-  float* dn_out;
+  uint16_t* gu_out;
+  uint16_t* dn_out;
   // Carve the regions out of a buffer of at least RequiredBytes().
   void Init(uint8_t* base);
 };
